Fix lint

+++ operator/scripts/run_update_golden_snapshots.sh
@@ -26,10 +26,9 @@ SCRIPTPATH="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
 ROOTDIR="${SCRIPTPATH}/../.."
 
 MANIFESTS_DIR="${ROOTDIR}/manifests"
-PROFILES_DIR="${ROOTDIR}/operator/data/profiles"
 CHARTS_SNAPSHOT="${ROOTDIR}/operator/cmd/mesh/testdata/manifest-generate/data-snapshot"
 
 # Clean up existing files
-rm -Rf "${CHARTS_SNAPSHOT}/"
+rm -Rf "${CHARTS_SNAPSHOT:?}/"
 
 cp -Rf "${MANIFESTS_DIR}" "${CHARTS_SNAPSHOT}"/
